Name the layout tests after what they now assert
Two test names outlived the behaviour they described. "Only the
all-panels layout shows the inspector" and "editor-only hides the list,
the other layouts show it" were both true until the fourth case, and
both kept passing after it because each enumerated the cases by hand
instead of iterating allCases. Green suites, stale claims. Both are now
driven from allCases, so a fifth case cannot slip past them the same way.

The shortcut mapping was missing its fourth arm. everyLayoutHasAShortcut
proves 4 resolves to something, but nothing pinned it to
editorAndInspector, so swapping the 3 and 4 arms would have passed. The
pairings are now parameterised, with a count check so a case added
without a pairing fails rather than going unnoticed.

The layout suite also moves out of DocumentIconTests.swift, which it had
outgrown and never belonged in, into EditorLayoutModeTests.swift.

And effectiveWidth's comment claimed the chosen width always survives.
True of a window resize, which is what the test covers; not true of a
drag in a shrunken window, which re-chooses from where the handle is and
persists that. The comment now says which is which.
